// Del 5/Angular/AdminScreens/src/app/model/ArtClasses/art-class.ts
export class ArtClass {
    artClassID: number;
    artClassName: string;
    artClassDescription: string;
    artClassStartDate: string;
    artClassEndDate: string;
    artClassLimit: string;
    refundDayLimit: string;
    classPrice: string;

    artClassType: string;

    venue: string;

    classTeacher: string;

    artClassAnnouncement: string;

    organisation: string;
}